Implement Accounts Model + Migrations + etc.
-Hardcode default Account for all Transactions (temp)
-Change double variables to decimal
-Minor formatting updates

+++ FinanceApp.Data/Requests/Transaction/Create.cs
@@ -8,7 +8,7 @@ public class Create : ISingleTransaction
 {
     public static string Flag => "<CreateTransaction>";
 
-    public required RequestField<double> Value { get; init; }
+    public required RequestField<decimal> Value { get; init; }
     public required RequestField<Models.Counterparty> Counterparty { get; init; }
 
     public override string ToString()
