+++ FinanceApp.MauiClient/ViewModel/QuickAddViewModel.cs
@@ -13,7 +13,7 @@ namespace FinanceApp.MauiClient.ViewModel;
 public partial class QuickAddViewModel(ServerConnection serverConnection) : BaseViewModel(serverConnection)
 {
 	[ObservableProperty]
-	private double _value;
+	private decimal _value;
     [ObservableProperty]
     private string _valueError = string.Empty;
